# docs/labos-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vyoma LabOS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reports/{id}/download": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download Report Pdf",
        "description": "Securely download/view PDF file for a report. Enforces tenant isolation. Audits download action.",
        "operationId": "download_report_pdf_api_v1_reports__id__download_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{id}/metadata": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report Metadata",
        "description": "Get detailed report metadata including verification details, file hashes, and download links.",
        "operationId": "get_report_metadata_api_v1_reports__id__metadata_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Report Metadata Api V1 Reports  Id  Metadata Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{id}/results": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Verified Report Results",
        "description": "Get structured JSON verification results data for clinical analysis & M2M orchestrators.",
        "operationId": "get_verified_report_results_api_v1_reports__id__results_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Verified Report Results Api V1 Reports  Id  Results Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/integrations": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Get Integration Status",
        "description": "Get n8n integration status and delivery counters for current organization. Admin only.",
        "operationId": "get_integration_status_api_v1_integrations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/integrations/n8n/test": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Test N8N Connection",
        "description": "Triggers safe test event (integration.test) to n8n webhook URL. Admin only. No patient data sent.",
        "operationId": "test_n8n_connection_api_v1_integrations_n8n_test_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationTestResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/integrations/logs": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Get Integration Logs",
        "description": "Paginated audit log of integration delivery attempts. Admin only. Tenant isolated.",
        "operationId": "get_integration_logs_api_v1_integrations_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (Pending, Sent, Failed)",
              "title": "Status"
            },
            "description": "Filter by status (Pending, Sent, Failed)"
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event_type",
              "title": "Event Type"
            },
            "description": "Filter by event_type"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationDeliveryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/logs/{id}/retry": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Retry Integration Delivery",
        "description": "Manually retry a failed integration delivery. Admin only. Enforces tenant isolation.",
        "operationId": "retry_integration_delivery_api_v1_integrations_logs__id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationDeliveryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/reports/{id}/download": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "M2M Download Report Pdf",
        "description": "Machine-to-Machine (M2M) secure endpoint for n8n to download report PDF.\nRequires valid X-Integration-Key header matching N8N_INTEGRATION_KEY.\nEnforces tenant isolation and audits M2M access.",
        "operationId": "m2m_download_report_pdf_api_v1_integrations_reports__id__download_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "X-Integration-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Integration-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/reports/{id}/metadata": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "M2M Get Report Metadata",
        "description": "Machine-to-Machine (M2M) endpoint for n8n/AI orchestrators to fetch report metadata.",
        "operationId": "m2m_get_report_metadata_api_v1_integrations_reports__id__metadata_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "X-Integration-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Integration-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response M2M Get Report Metadata Api V1 Integrations Reports  Id  Metadata Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/reports/{id}/results": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "M2M Get Verified Report Results",
        "description": "Machine-to-Machine (M2M) endpoint for n8n/AI orchestrators to fetch verified lab test results.",
        "operationId": "m2m_get_verified_report_results_api_v1_integrations_reports__id__results_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "X-Integration-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Integration-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response M2M Get Verified Report Results Api V1 Integrations Reports  Id  Results Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntegrationDeliveryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IntegrationDeliveryResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "IntegrationDeliveryListResponse"
      },
      "IntegrationDeliveryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "response_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "event_id",
          "event_type",
          "destination",
          "status",
          "attempts",
          "created_at",
          "updated_at"
        ],
        "title": "IntegrationDeliveryResponse"
      },
      "IntegrationStatusResponse": {
        "properties": {
          "is_configured": {
            "type": "boolean",
            "title": "Is Configured"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "sent_count": {
            "type": "integer",
            "title": "Sent Count"
          },
          "pending_count": {
            "type": "integer",
            "title": "Pending Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "last_successful_delivery": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Successful Delivery"
          },
          "last_failed_delivery": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Failed Delivery"
          }
        },
        "type": "object",
        "required": [
          "is_configured",
          "status",
          "sent_count",
          "pending_count",
          "failed_count"
        ],
        "title": "IntegrationStatusResponse"
      },
      "IntegrationTestResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "event_id",
          "message"
        ],
        "title": "IntegrationTestResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}